Move prelude & wrapper insertion to link
