+++ packages/packagers/js/src/JSPackager.js
@@ -68,7 +68,7 @@ export default (new Packager({
     // If scope hoisting is enabled, we use a different code path.
     if (bundle.env.shouldScopeHoist) {
       let wrappedAssets = new Set<string>();
-      let {ast, referencedAssets} = link({
+      let ast = link({
         bundle,
         bundleGraph,
         ast: await concat({
@@ -76,7 +76,6 @@ export default (new Packager({
           bundleGraph,
           options,
           wrappedAssets,
-          parcelRequireName,
         }),
         options,
         wrappedAssets,
@@ -87,11 +86,8 @@ export default (new Packager({
       traverse.cache.clear();
 
       let {contents, map} = generate({
-        bundleGraph,
         bundle,
         ast,
-        referencedAssets,
-        parcelRequireName,
         options,
       });
       return replaceReferences({
